Alternative idea to custom webview client
